refactor: centralize avatar/activity row fixtures in service-test-utils
Consolidates the three parallel copies of avatar/activity row composites
and mock factories (active-avatar lib, service-avatar, service-activity)
into @vers/service-test-utils/bun, alongside createViewer, so no
duplicate fixture ever ships again.
