--- libs/data/active-avatar/src/test-utils/index.ts
@@ -1,5 +1,0 @@
-export { createMockActivityRow } from './factories/create-mock-activity-row';
-export { createMockAvatarRow } from './factories/create-mock-avatar-row';
-export { createActiveAvatarRow } from './create-active-avatar-row';
-export { createActivityRow } from './create-activity-row';
-export { createAvatarRow } from './create-avatar-row';
