+++ libs/testing/service-test-utils/src/bun/composites/create-active-avatar-row.ts
