+++ libs/testing/service-test-utils/src/bun/composites/create-avatar-row.test.ts
@@ -1,12 +1,13 @@
 import { expect, test } from 'bun:test';
-import { createTestDB, createTestUser } from '@vers/service-test-utils/bun';
+import { createTestDB } from '../create-test-db';
+import { createTestUser } from '../create-test-user';
 import { createAvatarRow } from './create-avatar-row';
 
 test('it inserts an avatar row owned by the given user', async () => {
   await using testDB = await createTestDB();
 
-  const user = await createTestUser(testDB.db);
-  const avatar = await createAvatarRow(testDB.db, { userId: user.user.id });
+  const created = await createTestUser(testDB.db);
+  const avatar = await createAvatarRow(testDB.db, { userId: created.user.id });
 
   const row = await testDB.db
     .selectFrom('avatars')
@@ -14,17 +15,17 @@ test('it inserts an avatar row owned by the given user', async () => {
     .where('id', '=', avatar.id)
     .executeTakeFirstOrThrow();
 
-  expect(row.userId).toBe(user.user.id);
+  expect(row.userId).toBe(created.user.id);
 });
 
 test('it applies overrides on top of the faker-generated defaults', async () => {
   await using testDB = await createTestDB();
 
-  const user = await createTestUser(testDB.db);
+  const created = await createTestUser(testDB.db);
 
   const avatar = await createAvatarRow(testDB.db, {
     name: 'Foreign',
-    userId: user.user.id,
+    userId: created.user.id,
   });
 
   expect(avatar).toMatchObject({ name: 'Foreign' });
