+++ libs/testing/service-test-utils/src/bun/composites/create-avatar-row.ts
@@ -1,20 +1,20 @@
 import type { Avatars, DB } from '@vers/db';
 import type { Insertable, Kysely, Selectable } from 'kysely';
-import { createMockAvatarRow } from './factories/create-mock-avatar-row';
+import { createMockAvatar } from '../factories/create-mock-avatar';
 
 interface CreateAvatarRowData extends Partial<Insertable<Avatars>> {
   readonly userId: string;
 }
 
 /**
- * Seeds an avatar owned by the given user — the live-run and selection queries under test need a
- * real owning avatar to query against.
+ * Seeds an avatar owned by the given user. A cross-tenant fixture is the same composite called
+ * with a different `userId`.
  */
 export function createAvatarRow(
   db: Kysely<DB>,
   data: Readonly<CreateAvatarRowData>,
 ): Promise<Selectable<Avatars>> {
-  const row = createMockAvatarRow(data);
+  const row = createMockAvatar(data);
 
   return db.insertInto('avatars').values(row).returningAll().executeTakeFirstOrThrow();
 }
