+++ libs/testing/service-test-utils/src/bun/factories/create-mock-avatar.ts
