+++ services/activity/src/handlers/get-latest-activity-progress.test.ts
@@ -2,6 +2,7 @@ import { expect, test } from 'bun:test';
 import type { ActivityContract } from '@vers/contract-activity';
 import {
   createAnonymousViewer,
+  createAvatarRow,
   createServiceToken,
   createTestDB,
   createViewer,
@@ -10,7 +11,6 @@ import {
 import { createSimVersionRow } from '@vers/sim-registry/test-utils';
 import { buildRPCTestClient } from '@vers/test-utils';
 import { createActivityService } from '../create-activity-service';
-import { createAvatarRow } from '../test-utils/create-avatar-row';
 import { createMockCheckpointBatch } from '../test-utils/factories/create-mock-checkpoint-batch';
 
 /**
