+++ services/activity/src/handlers/stop-activity.test.ts
@@ -3,6 +3,7 @@ import type { ActivityContract } from '@vers/contract-activity';
 import { mockReplayService } from '@vers/mock-services/replay';
 import {
   createAnonymousViewer,
+  createAvatarRow,
   createServiceToken,
   createTestDB,
   createViewer,
@@ -12,7 +13,6 @@ import { createSimVersionRow } from '@vers/sim-registry/test-utils';
 import { buildRPCTestClient, waitFor } from '@vers/test-utils';
 import { createActivityService } from '../create-activity-service';
 import { server } from '../mocks/server';
-import { createAvatarRow } from '../test-utils/create-avatar-row';
 import { createMockCheckpointBatch } from '../test-utils/factories/create-mock-checkpoint-batch';
 
 /**
