+++ services/activity/src/handlers/track-activity-progress.test.ts
@@ -3,6 +3,7 @@ import type { ActivityContract } from '@vers/contract-activity';
 import { buildFailureXPLoss } from '@vers/idle-core';
 import {
   createAnonymousViewer,
+  createAvatarRow,
   createServiceToken,
   createTestDB,
   createViewer,
@@ -11,7 +12,6 @@ import {
 import { createSimVersionRow } from '@vers/sim-registry/test-utils';
 import { buildRPCTestClient } from '@vers/test-utils';
 import { createActivityService } from '../create-activity-service';
-import { createAvatarRow } from '../test-utils/create-avatar-row';
 import { createMockCheckpointBatch } from '../test-utils/factories/create-mock-checkpoint-batch';
 
 /**
