--- services/avatar/src/test-utils/create-avatar-row.ts
@@ -1,20 +1,0 @@
-import type { Avatars, DB } from '@vers/db';
-import type { Insertable, Kysely, Selectable } from 'kysely';
-import { createMockAvatar } from './factories/create-mock-avatar';
-
-interface CreateAvatarRowData extends Partial<Insertable<Avatars>> {
-  readonly userId: string;
-}
-
-/**
- * Seeds an avatar owned by *another* user for cross-tenant ownership tests, which can't go through
- * the owner-scoped RPC client.
- */
-export function createAvatarRow(
-  db: Kysely<DB>,
-  data: Readonly<CreateAvatarRowData>,
-): Promise<Selectable<Avatars>> {
-  const row = createMockAvatar(data);
-
-  return db.insertInto('avatars').values(row).returningAll().executeTakeFirstOrThrow();
-}
